Added repository scanner and empty test runners
We scan a directory ($HOME/repositories by default) looking for projects
to smoke.

Each projects is inside a direcotory. If such directory has a .smoke shell
script or a 't/' directory, we use that to smoke the project.

If now, we scan the project directory looking for subprojects in the same
way: if any subdirectory has a '.smoke' or a 't/' dir, we smoke it.

Signed-off-by: Pedro Melo <melo@simplicidade.org>